rename big to bigplanar
